feat: allow debugging of errors in production (#8502)
