+++ packages/central-server/app/middleware/errorHandler.js
@@ -1,7 +1,5 @@
-import { BaseError as SequelizeError } from 'sequelize';
 import { convertDatabaseError } from '@tamanu/database';
-import { Problem } from '@tamanu/errors';
-import { log } from '@tamanu/shared/services/logging';
+import { errorHandlerProblem } from '@tamanu/shared/utils';
 
 // eslint-disable-next-line no-unused-vars
 export const buildErrorHandler = getResponse => (error, req, res, next) => {
@@ -11,32 +9,10 @@ export const buildErrorHandler = getResponse => (error, req, res, next) => {
     return;
   }
 
-  if (error instanceof SequelizeError) {
-    error = convertDatabaseError(error);
-  }
-
-  const problem = (
-    error instanceof Problem ? error : Problem.fromError(error)
-  ).excludeSensitiveFields(process.env.NODE_ENV === 'production');
-
-  if (problem.status >= 500) {
-    log.error(`Error ${problem.status} (${problem.type}): `, error);
-  } else {
-    log.info(`Error ${problem.status} (${problem.type}): `, error);
-  }
+  const { problem, json } = errorHandlerProblem(error, req, { convertDatabaseError });
 
   res.set(problem.headers);
-  res.status(problem.status).send(getResponse(error, problem.toJSON()));
+  res.status(problem.status).send(getResponse(error, json));
 };
 
-export const defaultErrorHandler = buildErrorHandler((error, problem) => ({
-  // RFC 7807 Problem Details for HTTP APIs
-  ...problem,
-
-  // legacy error format
-  error: {
-    message: error.message,
-    name: error.name,
-    status: problem.status,
-  },
-}));
+export const defaultErrorHandler = buildErrorHandler((_error, defaultJson) => defaultJson);
