+++ packages/facility-server/app/middleware/errorHandler.js
@@ -1,25 +1,14 @@
-import { BaseError as SequelizeError } from 'sequelize';
 import { convertDatabaseError } from '@tamanu/database';
-import { Problem } from '@tamanu/errors';
-import { log } from '@tamanu/shared/services/logging';
+import { errorHandlerProblem } from '@tamanu/shared/utils';
 
-// eslint-disable-next-line no-unused-vars
-export default function errorHandler(error, req, res, _) {
-  if (error instanceof SequelizeError) {
-    error = convertDatabaseError(error);
+export default function errorHandler(error, req, res, next) {
+  // see https://expressjs.com/en/guide/error-handling.html#the-default-error-handler
+  if (res.headersSent) {
+    next(error);
+    return;
   }
 
-  const problem = (
-    error instanceof Problem ? error : Problem.fromError(error)
-  ).excludeSensitiveFields(process.env.NODE_ENV === 'production');
-
-  if (problem.type.includes('auth')) {
-    log.warn(`Error ${problem.status} (${problem.type}): ${error.message}`);
-  } else if (problem.status >= 500) {
-    log.error(`Error ${problem.status} (${problem.type}): `, error);
-  } else {
-    log.info(`Error ${problem.status} (${problem.type}): `, error);
-  }
+  const { problem, json } = errorHandlerProblem(error, req, { convertDatabaseError });
 
   // we're past the point of permission checking; this just
   // makes sure the error send doesn't get intercepted by the
@@ -29,11 +18,5 @@ export default function errorHandler(error, req, res, _) {
   }
 
   res.set(problem.headers);
-  res.status(problem.status).send({
-    ...problem.toJSON(),
-    error: {
-      message: error.message,
-      ...error,
-    },
-  });
+  res.status(problem.status).send(json);
 }
