+++ packages/upgrade/package.json
@@ -28,13 +28,14 @@
     "clean": "rimraf dist *.tsbuildInfo",
     "clean:deps": "rimraf node_modules",
     "test": "vitest run --no-file-parallelism",
-    "package-default-translations": "npx tsx scripts/scrapeTranslations.ts > dist/default-translations.json",
+    "package-default-translations": "tsx scripts/scrapeTranslations.ts > dist/default-translations.json",
     "test:watch": "vitest watch"
   },
   "devDependencies": {
     "@types/node": "^20.9.0",
     "@types/toposort": "^2.0.7",
     "rimraf": "^6.0.1",
+    "tsx": "^4.20.6",
     "type-fest": "^4.31.0",
     "vitest": "^2.1.9"
   },
